+++ src/components/SectionTitle/styled.ts
